+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 16:16:53 by alechin           #+#    #+#             */
-/*   Updated: 2025/07/01 16:18:51 by alechin          ###   ########.fr       */
+/*   Updated: 2025/07/02 14:41:46 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,19 +25,59 @@
 # include "../libft/ft_printf/ft_printf.h"
 # include "../libft/Wglbx/wglbx.h"
 
-# include "execution.h"
-# include "parsing.h"
-
 # define SHELL_NAME "bass shell"
 # define STDIN 0
 # define STDOUT 1
 # define STDERR 2
 
-# define PIPE_LEFT 0
-# define PIPE_RIGHT 1
-
 # define SUCCESS 0
 # define FAILURE 1
+
+typedef enum s_token_type
+{
+	TOKEN_WORD,
+	TOKEN_PIPE,
+	TOKEN_REDIRECT_IN,// <
+	TOKEN_REDIRECT_OUT,// >
+	TOKEN_REDIRECT_APPEND,// >>
+	TOKEN_HEREDOC,// <<
+	TOKEN_EOF,
+	TOKEN_ERROR
+}	t_token_type;
+
+typedef struct s_token
+{
+	t_token_type	type;
+	char			*value;
+	bool			has_expansion;/*marks if token contains $ for expansion*/
+	struct s_token	*next;
+}	t_token;
+
+typedef struct s_env
+{
+	char			*key;
+	char			*value;
+	struct s_env	*next;
+}	t_env;
+
+typedef struct s_redir
+{
+	int				type;
+	char			*file;
+	struct s_redir	*next;
+}	t_redir;
+
+typedef struct s_exec
+{
+	char			**argv;
+	char			*cmd_path;
+	int				infile;
+	int				outfile;
+	bool			has_pipe;
+	bool			is_builtin;
+	t_redir			*redirs;
+	struct s_exec	*next;
+}	t_exec;
 
 typedef enum e_status
 {
@@ -45,16 +85,6 @@ typedef enum e_status
 	QUOTE,
 	DOUBLE_QUOTE,
 }	t_status;
-
-typedef struct s_root
-{
-	int				level;
-	t_token			**tokens;
-	t_minishell		*msh;
-	struct s_root	*origin;
-	struct s_root	*left;
-	struct s_root	*right;
-}	t_root;
 
 typedef struct s_minishell
 {
@@ -65,6 +95,16 @@ typedef struct s_minishell
 	t_exec			*cmds;
 	struct s_base	*base;
 }	t_minishell;
+
+typedef struct s_root
+{
+	int				level;
+	t_token			**tokens;
+	t_minishell		*msh;
+	struct s_root	*origin;
+	struct s_root	*left;
+	struct s_root	*right;
+}	t_root;
 
 typedef struct s_base
 {
@@ -77,82 +117,38 @@ typedef struct s_base
 	struct s_base	*prev;
 }	t_base;
 
-/* -- Error Handling -- */
-int 	error2exit(char *msg, int status);
-int 	error4exit(char *msg, int status);
-int 	error6exit(char *msg, int status);
+typedef struct s_lexer
+{
+	char		*input;
+	int			pos;
+	char		current_char;
+	bool		in_single_quote;
+	bool		in_double_quote;
+}	t_lexer;
 
-/* -- Built-ins -- */
-int		koi_cd(char **path);
-int		koi_echo(char **cmd);
-int		koi_env(void);
-int		koi_exit(char **cmd, t_root *root);
-int		koi_export(char **cmd);
-int		koi_pwd(void);
-int		koi_unset(char **cmd, t_root *root);
+typedef enum s_node_type
+{
+	NODE_COMMAND,
+	NODE_PIPE,
+	NODE_REDIRECT_IN,
+	NODE_REDIRECT_OUT,
+	NODE_REDIRECT_APPEND,
+	NODE_HEREDOC
+}	t_node_type;
 
-/* -- Execution Built-ins -- */
-int		specify(char **cmd);
-int		is_builtin(char **cmd, t_minishell *std);
+typedef struct s_ast_node
+{
+	t_node_type			type;
+	char				**args;
+	char				*filename;
+	struct s_ast_node	*left;
+	struct s_ast_node	*right;
+}	t_ast_node;
 
-/* -- Export Helper -- */
-void	no_args(t_minishell *e);
-int		valid_name(char *env, t_minishell *e);
-int		valid_environment(char **env, t_minishell *e);
-char	**appends(char **env, t_minishell *e, int not_equals);
-
-/* -- Branching -- */
-int		is_fork(t_root *root);
-int		execute_status(t_root *root);
-void	dup2io(int io_in, int io_out);
-
-/* -- Execution Main -- */
-void	close_fds(void);
-char	type_null(t_root *root);
-int		execution(t_root *root);
-
-/* -- Execution Path -- */
-char	*exec_path(char *cmd, char **env);
-
-/* -- Execution External -- */
-int		external(char **cmd, t_minishell *msh);
-
-/* -- Pipe -- */
-int		pipe(char **cmd, char n, t_root *root);
-
-/* -- Redirections Exec -- */
-int 	redir_type(int type);
-int		redirect_prompt(t_root *root);
-int		redirection(t_root *root);
-
-/* -- Redirections Operations -- */
-int		in(char *filename);
-int		out(char *filename);
-int		append(char *filename);
-int		heredoc(char *filename, char *limiter);
-
-/* -- Expansion Main -- */
-char	*remove_quotes(char *str);
-char	**join_commands(t_root *root);
-char	**expand_commands(char **cmd, t_root *root);
-char	*expand_string(char *str);
-
-/* -- Expansion Helper -- */
-char	**remainder(char **cmd, t_root *root, int box, int keep);
-char	*remove_quotes_size(char *str, char *i);
-char	**remove_null(char **cmd, t_root *root);
-char	*get_raw_area(char *str, int *i);
-char	*get_next_area(char *str, int *i);
-
-/* -- Env -- */
-int		variable_len(char *start);
-char	*to_get_env(char *start, int len);
-
-/* -- Dollar -- */
-char	*expand_dollar(char *prompt);
-
-/* -- Termination -- */
-void	terminate_ast(t_root **root);
-void	terminate_tokens(t_token **tokens);
+typedef struct s_parser
+{
+	t_token	*tokens;
+	t_token	*current_token;
+}	t_parser;
 
 #endif
