+++ lexer/remove_quotes.c
