+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 12:08:17 by alechin           #+#    #+#             */
-/*   Updated: 2025/07/01 16:20:09 by alechin          ###   ########.fr       */
+/*   Updated: 2025/07/01 16:21:23 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
