+++ main/main.c
@@ -6,7 +6,7 @@
 /*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/23 13:59:16 by alechin           #+#    #+#             */
-/*   Updated: 2025/06/26 10:25:48 by alechin          ###   ########.fr       */
+/*   Updated: 2025/07/02 15:39:16 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,9 @@ int	main(int argc, char **argv, char **envp)
 	int		status;
 
 	status = NORMAL;
-	while (1)
-	{
-		cmd = readline("\033[33m-- BASS AMATEUR SHELL --
-			\033[36m\n[🐡 FISH BITES] o->\033[0m ");
-	}
+	setup_signals();
+	setup_history();
+	shell_loop();
+	cleanup_history();
 	return (0);
 }
