+++ misc/signalling.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signalling.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/30 22:09:03 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/07/01 00:45:56 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:49 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
