+++ parser/freeing.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   freeing.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/17 14:00:58 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/30 22:02:00 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:30 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
