+++ parser/initializing.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   initializing.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/11 23:18:38 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/29 21:20:55 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:30 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
