--- parser/main.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   main.c                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/06/30 22:31:54 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/30 22:32:27 by rpadasia         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../includes/parsing.h"
-
-int	main(void)
-{
-	setup_signals();
-	setup_history();
-	shell_loop();
-	cleanup_history();
-	return (0);
-}
