+++ parser/next.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   next.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/11 23:26:42 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/29 21:35:35 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:29 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
