+++ parser/parse_command.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_command.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/18 21:32:17 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/29 21:36:22 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:28 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
