+++ parser/parse_pipes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_pipes.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/17 14:20:48 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/29 21:43:47 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:27 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
