+++ parser/parse_redirections.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_redirections.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpadasia <ryanpadasian@gmail.com>          +#+  +:+       +#+        */
+/*   By: alechin <alechin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/19 15:29:56 by rpadasia          #+#    #+#             */
-/*   Updated: 2025/06/29 21:45:24 by rpadasia         ###   ########.fr       */
+/*   Updated: 2025/07/02 09:59:27 by alechin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
