+++ printtemp.c
