--- signal/gilbert.c
@@ -1,8 +1,0 @@
-/*			Please Delete This			*/
-
-#include <stdio.h>
-void gilbert()
-{
-	printf("gilbert\n");
-	return ;
-}
