--- wildcard/wildcard.c
@@ -1,9 +1,0 @@
-#ifndef WILDCARD
-# define WILDCARD "who's doing this???"
-
-void ft_print_wildcard(char *message)
-{
-	/* do delete this when we actually get there */
-}
-
-# endif
